+++ iotdb-0.09/src/main/java/cn/edu/tsinghua/iotdb/benchmark/iotdb009/IoTDBSession.java
@@ -1,4 +1,4 @@
-package cn.edu.tsinghua.iotdb.benchmark.iotdb010;
+package cn.edu.tsinghua.iotdb.benchmark.iotdb009;
 
 import cn.edu.tsinghua.iotdb.benchmark.conf.Config;
 import cn.edu.tsinghua.iotdb.benchmark.conf.ConfigDescriptor;
